commit c289b5ff004ff05fc034f4dfd0b9e3a29c17b673
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Mon Jul 11 02:37:45 2022 +0900

    Refactor

commit 3dc8cf09f9af965c778103ed95d03c49f2dbc8ee
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Mon Jul 11 02:30:28 2022 +0900

    Refactor

commit 9c21a9270869ad1be2e38dd53d321e60d04db7fb
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Mon Jul 11 01:58:19 2022 +0900

    Refactor

commit c91166bc8e158c968c406d5afddebacabf0c60da
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Mon Jul 11 00:38:34 2022 +0900

    Refactor

commit db16495a164b8c399bd19091933c128bdfc1f6ff
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Sun Jul 10 14:17:29 2022 +0900

    Refactor

commit a76d8f39e2ccb958c794dd40caeb3a88f4115ed9
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Sun Jul 10 13:35:46 2022 +0900

    Fix

commit 6526c2e4d03b54c69fd11ced1ca582e8f3220fbf
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Mon Jul 4 09:05:07 2022 +0900

    Add TextInput

commit dcbe856702909947fa9099b088ba15a403d75399
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Mon Jul 4 02:04:34 2022 +0900

    Add component

commit 066d29bdbd5436d2ecf44a28cec930b33520d5e5
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Sun Jul 3 19:34:25 2022 +0900

    Refactor

commit 2540ab9780ff6d4801441109b920c5454d8d0d2d
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Sun Jul 3 19:24:47 2022 +0900

    Add text decorations

commit 38e034c6bb0dc70118397d8dab63982207164f64
Author: dytlzl <dytlzl.dev@gmail.com>
Date:   Sat Jul 2 23:08:29 2022 +0900

    Fix

+++ example/cmd/colorcode/main.go
@@ -16,7 +16,7 @@ func rootView() *tui.View {
 		return tui.InlineStack(
 			tui.InlineMapN(16, func(j int) *tui.View {
 				seq := i*16 + j
-				return tui.Fmt("%4d", seq).FGColor(seq)
+				return tui.Fmt("%4d", seq).FGColor(uint8(seq))
 			}),
 			tui.Break(),
 		)
